Add path normalization middleware
I realized half-way through implementing this that this is
already being implemented in KeepSafe/aiohttp#362, so right
now this is mostly a copy-n-paste of that code.
Keeping this for now because it is still useful to have the
setup function for reading app config.

Once that PR is merged, much of this code can go away.